Moved function to namespace, restructured how indexes are located.
